+++ cm-mlops/script/process-ae-users/customize.py
@@ -1,0 +1,10 @@
+from cmind import utils
+import cmind as cm
+import os
+
+def preprocess(i):
+
+    os_info = i['os_info']
+    env = i['env']
+
+    return {'return':0}